# Speed up `--all-features` build by 33%~55% by removing cfg