--- a/build/src/fs/cargo_toml.rs
+++ b/build/src/fs/cargo_toml.rs
@@ -20,18 +20,10 @@ impl CargoToml {
                     short_name: &'a str,
                     crate_version: String,
                     icon_package_name: &'a str,
-                    features: Vec<&'a str>,
                 }
-
-                let pkg_icons = pkg.icons();
-                let features = pkg_icons
-                    .iter()
-                    .map(|icon| &*icon.feature.name)
-                    .collect::<Vec<_>>();
 
                 Ok(Template {
                     crate_version: pkg.meta.crate_version.to_string(),
-                    features,
                     short_name: &pkg.meta.short_name,
                     icon_package_name: &pkg.meta.package_name,
                 }
@@ -62,22 +54,15 @@ impl CargoToml {
             LibType::IconIndex => {
                 #[derive(Template)]
                 #[template(path = "icon_index/Cargo.toml", escape = "none")]
-                struct Template<'a> {
-                    features: Vec<&'a str>,
-                }
+                struct Template {}
 
-                let features = crate::Packages::get_icons()?
-                    .map(|icon| &*icon.feature.name)
-                    .collect::<Vec<_>>();
-
-                Ok(Template { features }.render()?)
+                Ok(Template {}.render()?)
             }
 
             LibType::Boilerplate => {
                 #[derive(Template)]
                 #[template(path = "boilerplate/Cargo.toml", escape = "none")]
                 struct Template<'a> {
-                    features: Vec<&'a str>,
                     short_names: Vec<&'a str>,
                 }
 
@@ -85,13 +70,8 @@ impl CargoToml {
                     .iter()
                     .map(|pkg| pkg.meta.short_name.as_ref())
                     .collect::<Vec<_>>();
-                let features = vec![];
 
-                Ok(Template {
-                    short_names,
-                    features,
-                }
-                .render()?)
+                Ok(Template { short_names }.render()?)
             }
         }
     }