--- a/build/templates/icon_index/Cargo.toml
+++ b/build/templates/icon_index/Cargo.toml
@@ -32,7 +32,4 @@ once_cell = "1.17"
 
 leptos_icons = { version = "0.0.13", default_features = false, features = [
 "strum",
-{% for feat in features %}
-"{{ feat }}",
-{%- endfor %}
 ] }