--- a/icondata/Cargo.toml
+++ b/icondata/Cargo.toml
@@ -15,24 +15,24 @@ rust-version = "1.60.0"
 icondata_core = "0.0.2"
 serde = { workspace = true, optional = true }
 
-icondata_ai = { version = "0.0.8", optional = true }
-icondata_bi = { version = "0.0.8", optional = true }
-icondata_bs = { version = "0.0.8", optional = true }
-icondata_cg = { version = "0.0.8", optional = true }
-icondata_ch = { version = "0.0.8", optional = true }
-icondata_fa = { version = "0.0.8", optional = true }
-icondata_fi = { version = "0.0.8", optional = true }
-icondata_hi = { version = "0.0.8", optional = true }
-icondata_im = { version = "0.0.8", optional = true }
-icondata_io = { version = "0.0.8", optional = true }
-icondata_lu = { version = "0.0.8", optional = true }
-icondata_oc = { version = "0.0.8", optional = true }
-icondata_ri = { version = "0.0.8", optional = true }
-icondata_si = { version = "0.0.8", optional = true }
-icondata_tb = { version = "0.0.8", optional = true }
-icondata_ti = { version = "0.0.8", optional = true }
-icondata_vs = { version = "0.0.8", optional = true }
-icondata_wi = { version = "0.0.8", optional = true }
+icondata_ai = { path = "../icondata_ai", version = "0.0.8", optional = true }
+icondata_bi = { path = "../icondata_bi", version = "0.0.8", optional = true }
+icondata_bs = { path = "../icondata_bs", version = "0.0.8", optional = true }
+icondata_cg = { path = "../icondata_cg", version = "0.0.8", optional = true }
+icondata_ch = { path = "../icondata_ch", version = "0.0.8", optional = true }
+icondata_fa = { path = "../icondata_fa", version = "0.0.8", optional = true }
+icondata_fi = { path = "../icondata_fi", version = "0.0.8", optional = true }
+icondata_hi = { path = "../icondata_hi", version = "0.0.8", optional = true }
+icondata_im = { path = "../icondata_im", version = "0.0.8", optional = true }
+icondata_io = { path = "../icondata_io", version = "0.0.8", optional = true }
+icondata_lu = { path = "../icondata_lu", version = "0.0.8", optional = true }
+icondata_oc = { path = "../icondata_oc", version = "0.0.8", optional = true }
+icondata_ri = { path = "../icondata_ri", version = "0.0.8", optional = true }
+icondata_si = { path = "../icondata_si", version = "0.0.8", optional = true }
+icondata_tb = { path = "../icondata_tb", version = "0.0.8", optional = true }
+icondata_ti = { path = "../icondata_ti", version = "0.0.8", optional = true }
+icondata_vs = { path = "../icondata_vs", version = "0.0.8", optional = true }
+icondata_wi = { path = "../icondata_wi", version = "0.0.8", optional = true }
 [features]
 strum = [
 "icondata_ai/strum",
@@ -76,21 +76,21 @@ serde = [
 "icondata_wi/serde",
 ]
 
-Ai = []
-Bi = []
-Bs = []
-Cg = []
-Ch = []
-Fa = []
-Fi = []
-Hi = []
-Im = []
-Io = []
-Lu = []
-Oc = []
-Ri = []
-Si = []
-Tb = []
-Ti = []
-Vs = []
-Wi = []
+Ai = ["icondata_ai"]
+Bi = ["icondata_bi"]
+Bs = ["icondata_bs"]
+Cg = ["icondata_cg"]
+Ch = ["icondata_ch"]
+Fa = ["icondata_fa"]
+Fi = ["icondata_fi"]
+Hi = ["icondata_hi"]
+Im = ["icondata_im"]
+Io = ["icondata_io"]
+Lu = ["icondata_lu"]
+Oc = ["icondata_oc"]
+Ri = ["icondata_ri"]
+Si = ["icondata_si"]
+Tb = ["icondata_tb"]
+Ti = ["icondata_ti"]
+Vs = ["icondata_vs"]
+Wi = ["icondata_wi"]