--- a/icondata_ai/README.md
+++ b/icondata_ai/README.md
@@ -1,6 +1,6 @@
 # Icondata - ai
 
-Icon data from the *Ant Design Icons* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *Ant Design Icons* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.