--- a/icondata_bi/README.md
+++ b/icondata_bi/README.md
@@ -1,6 +1,6 @@
 # Icondata - bi
 
-Icon data from the *BoxIcons* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *BoxIcons* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.