--- a/icondata_bs/README.md
+++ b/icondata_bs/README.md
@@ -1,6 +1,6 @@
 # Icondata - bs
 
-Icon data from the *Bootstrap Icons* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *Bootstrap Icons* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.