--- a/icondata_cg/README.md
+++ b/icondata_cg/README.md
@@ -1,6 +1,6 @@
 # Icondata - cg
 
-Icon data from the *css.gg* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *css.gg* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.