--- a/icondata_ch/README.md
+++ b/icondata_ch/README.md
@@ -1,6 +1,6 @@
 # Icondata - ch
 
-Icon data from the *Charm* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *Charm* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.