--- a/icondata_fa/README.md
+++ b/icondata_fa/README.md
@@ -1,6 +1,6 @@
 # Icondata - fa
 
-Icon data from the *Font Awesome* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *Font Awesome* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.