--- a/icondata_hi/README.md
+++ b/icondata_hi/README.md
@@ -1,6 +1,6 @@
 # Icondata - hi
 
-Icon data from the *Heroicons* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *Heroicons* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.