--- a/icondata_im/README.md
+++ b/icondata_im/README.md
@@ -1,6 +1,6 @@
 # Icondata - im
 
-Icon data from the *IcoMoon Free* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *IcoMoon Free* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.