--- a/icondata_io/README.md
+++ b/icondata_io/README.md
@@ -1,6 +1,6 @@
 # Icondata - io
 
-Icon data from the *Ionicons* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *Ionicons* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.