--- a/icondata_lu/README.md
+++ b/icondata_lu/README.md
@@ -1,6 +1,6 @@
 # Icondata - lu
 
-Icon data from the *Lucide* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *Lucide* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.