--- a/icondata_oc/README.md
+++ b/icondata_oc/README.md
@@ -1,6 +1,6 @@
 # Icondata - oc
 
-Icon data from the *Github Octicons* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *Github Octicons* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.