--- a/icondata_ri/README.md
+++ b/icondata_ri/README.md
@@ -1,6 +1,6 @@
 # Icondata - ri
 
-Icon data from the *Remix Icon* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *Remix Icon* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.