--- a/icondata_si/README.md
+++ b/icondata_si/README.md
@@ -1,6 +1,6 @@
 # Icondata - si
 
-Icon data from the *Simple Icons* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *Simple Icons* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.