--- a/icondata_tb/README.md
+++ b/icondata_tb/README.md
@@ -1,6 +1,6 @@
 # Icondata - tb
 
-Icon data from the *Tabler Icons* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *Tabler Icons* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.