--- a/icondata_ti/README.md
+++ b/icondata_ti/README.md
@@ -1,6 +1,6 @@
 # Icondata - ti
 
-Icon data from the *Typicons* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *Typicons* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.