--- a/icondata_vs/README.md
+++ b/icondata_vs/README.md
@@ -1,6 +1,6 @@
 # Icondata - vs
 
-Icon data from the *VS Code Icons* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *VS Code Icons* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.