--- a/icondata_wi/README.md
+++ b/icondata_wi/README.md
@@ -1,6 +1,6 @@
 # Icondata - wi
 
-Icon data from the *Weather Icons* library. Every icon is packaged as its own cargo feature to reduce build times.
+Icon data from the *Weather Icons* library.
 
 Visit this [repository](https://github.com/Carlosted/icondata) for in-depth information.