… Cargo.lock
… boilerplate/Cargo.toml
… icon_index/Cargo.toml
… icon_index/src/lib.rs
… icondata_ai/Cargo.toml
… icondata_ai/src/lib.rs
… icondata_bi/Cargo.toml
… icondata_bi/src/lib.rs
… icondata_bs/Cargo.toml
… icondata_bs/src/lib.rs
… icondata_cg/Cargo.toml
… icondata_cg/src/lib.rs
… icondata_ch/Cargo.toml
… icondata_ch/src/lib.rs
… icondata_fa/Cargo.toml
… icondata_fa/src/lib.rs
… icondata_fi/Cargo.toml
… icondata_fi/src/lib.rs
… icondata_hi/Cargo.toml
… icondata_hi/src/lib.rs
… icondata_im/Cargo.toml
… icondata_im/src/lib.rs
… icondata_io/Cargo.toml
… icondata_io/src/lib.rs
… icondata_lu/Cargo.toml
… icondata_lu/src/lib.rs
… icondata_oc/Cargo.toml
… icondata_oc/src/lib.rs
… icondata_ri/Cargo.toml
… icondata_ri/src/lib.rs