Currently, this library enables a `cfg(feature = "{name}")` for every single icon.

Unfortunately, feature resolution the way Cargo does it is not, and **cannot**, be "fast" beyond a certain point: it is at minimum a time cost that is linear with the number of features. At its worst, the package resolution problem is often equated with bool-SAT. In addition, this adds several thousands of lines of code to each crate which must be lexed, parsed, and so on. As I recall, adding a feature is also going to invalidate the build cache for a crate.

My personal machine is very fast, but this still makes the builds _notably_ faster.

| cargo...               | main   | remove-cfg-gen |
| ---------------------- | ------:| --------------:|
| `check --all-features` | ~14.5s |         ~11.5s |
| `build --all-features` | ~16.5s |         ~12.5s |

I believe almost everything is in order, here, but I found difficulty testing the result meaningfully. One of the libraries did not finish downloading so I backed out the generated diffs involving it.