@workingjubilee 2023-10-12
This has different effects on the zero-feature build case for this _repo_.

However, your dependents are going to be enabling features, or else this crate will be roughly equivalent to another Rust file that is also very fast to compile:
```rust

```
When they add a feature, they will discard the incremental build cache again on the next compiler run. In many cases, people will have a compiling daemon build their crate periodically, which, unless they type very fast, will result in discarding and rebuilding this crate once for each feature they add, if memory serves. After doing this even 100 times, I expect, this will make `node_modules` seem lightweight.

Meanwhile, each individual crate itself does not seem to suffer a notable impact from being built without cfg. This also affects the download of the crate and its index (which https://github.com/rust-lang/crates.io/issues/7269 mentions as containing **all** the features), which are also factors, especially in first-time addition of this crate or other fresh-slate builds.