@Carlosted 2023-10-12
Thank you for reaching out!

The first intentions with protecting every single icon behind a feature is to reduce `wasm` bundle sizes. I believe `web-sys` is also feature gating their structs for the same reason but I may be wrong 😅. After benchmarking on my machine, it seems that not using feature flags per icon and importing _only 1 icon crate_ (i.e., `icondata_fa`) bloats the bundle by __1.5MB__!!

I'm afraid we will have to find an alternative solution to resolve this issue. I am fully aware of the stress this creates on `Cargo` and `crates.io`, and I am willing to look into solutions if you can think of any.