@workingjubilee 2023-10-12
The web-sys crate is a direly undermaintained crate hacked together several years ago, when wasm tooling was much worse, yet expected to get better relatively quickly to the point of outmoding it in a few years, and then given up for other maintainers, who have since changed it further. There is no one who can meaningfully answer for its design choices anymore, so I will neither excuse them nor accuse them. I will say, however, that I think that any design choice that seems reasonable can become unreasonable if you add a few orders of magnitude to it.

From what I understand of how the compiler, linker, and other tools do dead code elimination, I believe that by creating a single type which references every single icon, the per-crate enum which then becomes an IconData, your crate's style virtually guarantees that any use of the enum with outside data that could make it take on any value would make dead code elimination almost impossible. That is assuming, of course, that the compilation tools adequately perform dead code elimination. It is plausible they do not.

If you wanted to assure that only used code is injected into the bundle, why not design your crate *as* a build tool, fundamentally? You're asking the user to list the icons they want, after all.