@workingjubilee 2023-10-12
I believe there are three good solutions:
- Work to significantly improve dead code elimination in general in the Rust/wasm ecosystem. That may be slightly blue-sky.
- Shift the build towards the final package, by asking programmers to provide a file that enumerates the list of icons they want, stowed in a variable Cargo can track, that then builds the data in, so that it is more a "just-in-time" phenomenon.
- Using `static`s instead of `const`s (so that the data is guaranteed to be in a single place), possibly combined with `#[link_section]` and possibly emitting metadata that makes it easier to track what gets actually used and put it into its own sections, and then DCE sections of statics quasi-manually using `wasm-strip` or the like.