@Carlosted 2023-10-13
> I believe there are three good solutions:
> 
> * Work to significantly improve dead code elimination in general in the Rust/wasm ecosystem. May be slightly blue-sky.

I believe I am highly underqualified for this job, and I will leave it to someone who is more familiar with rust and compiler architectures.

> * Shift the build towards the final package, by asking programmers to provide a file that enumerates the list of icons they want, stowed in a variable Cargo can track so that it can rerun-if-changed in the build.rs, that then builds the data in, so that it is more a "just-in-time" phenomenon.

This idea is for me the most compelling, and something I will try to achieve for the next release.

> * Using `static`s instead of `const`s (so that the data is guaranteed to be in a single place), possibly combined with `#[link_section]` and possibly emitting metadata that makes it easier to track what gets actually used and put it into its own sections, and then DCE sections of statics quasi-manually using `wasm-strip` or the like.

Similar to the first option, I do not know enough about Rust's ABI and `wasm` to undertake such a possibility.

I had another idea which I wanted to have your opinion on since you most certainly have more experience than me: _How about a proc-macro that has the whole icon data-table and that inlines the data at compile time?_

Thank you, your help means a lot!