@workingjubilee 2023-10-16
> I had another idea which I wanted to have your opinion on since you most certainly have more experience than me: How about a proc-macro that has the whole icon data-table and that inlines the data at compile time?

That seems reasonable. It's "morally equivalent" to the build.rs version, I think, but may have different implications regarding deduplication across multiple dependents.